+++ src/app/hooks/useModal/type.ts
@@ -1,0 +1,13 @@
+import { ReactNode } from "react";
+
+export interface UseModalResult {
+  Modal: ({ children, className }: ModalContentProps) => JSX.Element;
+  open: () => void;
+  close: () => void;
+  isOpen: boolean;
+}
+
+export interface ModalContentProps {
+  children: ReactNode;
+  className?: string;
+}
